refactor: Moved DiceRoller to it's own class and added logging.
