+++ Dice/IDiceRoller.cs
@@ -1,0 +1,4 @@
+public interface IDiceRoller
+{
+    (List<int> rolls, int total, int faces, int modifier) Roll(string diceExpr);
+}
